WIP: Checkpoint. About to extract the concept of DocumentLayout from editable_document.dart
